# api/openapi.yaml
openapi: 3.0.0
info:
  description: This is Debugging API document.
  version: 1.0.0
  title: Debugging REST API
  contact:
    name: Dasol Kim
    email: sorikiki98@naver.com
servers:
  - url: 'http://localhost:8080'
...
paths:
...
  /companies/interest/{company_id}:
    post:
      tags:
        - companies
      summary: Add a company into a user's company interest list
      description: Add one item into a scheme called CompanyInterest using id of both current user and chosed company.
      operationId: addCompanyInterest
      parameters:
        - name: company_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Company Interest Item Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InsertId'
        '409':
          description: Already Added With Same User and Company
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
    delete:
      tags:
        - companies
      summary: Remove a company from a user's company interest list
      description: Remove one item into a scheme called CompanyInterest using id of both current user and chosed company.
      operationId: removeCompanyInterest
      parameters:
        - name: company_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Company Interest Item Deleted
        '404':
          description: Unexisted Company Interest Item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
  /companies/reservation/{company_id}:
    post:
      tags:
        - companies
      summary: Make a reservation to the Debugging service
      description: Make a reservation using id of both current user and chosed company.
      operationId: reserve
      parameters:
        - name: company_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationApplication'
        required: true
      responses:
        '201':
          description: New Reservation Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InsertId'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
  /companies/reservation/{reservation_id}:
    get:
      tags:
        - companies
      summary: Fetch user's reservation information about a company
      description: Show a bunch of information including user, company and reservation.
      operationId: checkReservation
      parameters:
        - name: reservation_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successfully Fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationDetail'
        '404':
          description: Unexisted Reservation Item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
...
components:
  schemas:
...
    ReservationApplication:
      type: object
      properties:
        bugName:
          type: string
        firstFoundDate:
          type: string
          format: date
        firstFoundPlace:
          type: string
        wantedDate:
          type: string
        wantedTime:
          type: string
        hasBugBeenShown:
          type: boolean
        reserveDateTime:
          type: string
          format: date-time
        extraMessage:
          type: string
          nullable: true
      required:
        - bugName
        - firstFoundDate
        - firstFoundPlace
        - wantedDate
        - wantedTime
        - hasBugBeenShown
        - reserveDateTime
      example:
        bugName: 바퀴벌레
        firstFoundDate: '2022-05-04'
        firstFoundPlace: 화장실 변기 옆
        wantedDate: 2022-05-12~2022-05-19
        wantedTime: 오후 4시~ 오후 6시
        hasBugBeenShown: false
        reserveDateTime: '2022-05-10'
    ReservationDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
        bugName:
          type: string
        firstFoundDate:
          type: string
          format: date
        firstFoundPlace:
          type: string
        wantedDate:
          type: string
        wantedTime:
          type: string
        hasBugBeenShown:
          type: boolean
        reserveDateTime:
          type: string
          format: date-time
        extraMessage:
          type: string
          nullable: true
        userId:
          type: integer
          format: int64
        userName:
          type: string
        userContactNumbers:
          type: string
        userEmail:
          type: string
          format: email
        userAddress:
          type: string
          nullable: true
        sizeOfHouse:
          type: number
          nullable: true
        numOfRooms:
          type: integer
          nullable: true
        companyId:
          type: integer
          format: int64
        companyName:
          type: string
          maxLength: 12
        shortIntro:
          type: string
          maxLength: 30
        description:
          type: string
        companyContactNumbers:
          type: string
        killableBugs:
          type: string
        availableArea:
          type: string
        availableCounselTime:
          type: string
        thumbnail:
          type: string
          format: url
          nullable: true
        processState:
          type: integer
        engineerName:
          type: string
        engineerContactNumbers:
          type: string
        expectedEstimate:
          type: string
        visitDateTime:
          type: string
      required:
        - id
        - bugName
        - firstFoundDate
        - firstFoundPlace
        - wantedDate
        - wantedTime
        - hasBugBeenShown
        - userId
        - userName
        - userContactNumbers
        - userEmail
        - userAddress
        - companyId
        - companyName
        - shortIntro
        - description
        - companyContactNumbers
        - killableBugs
        - availableArea
        - availableCounselTime
      example:
        id: 1
        bugName: 바퀴벌레
        firstFoundDate: '2022-05-04'
        firstFoundPlace: 화장실 변기 옆
        wantedDate: 2022-05-12~2022-05-19
        wantedTime: 오후 4시~ 오후 6시
        hasBugBeenShown: false
        reserveDateTime: '2022-05-10'
        userId: 1
        userName: 김다솔
        userContactNumbers: 010-1111-2222
        userEmail: sorikiki98@sookmyung.ac.kr
        userAddress: 서울특별시 용산구 청파로
        companyId: 2
        companyName: 해충마스터
        shortIntro: 바퀴벌레, 좀벌레 등 해충방역 전문 업체입니다. 수도권만 출장 가능합니다.
        description: 해충마스터는 우리 인간에게 유해한 해충을 진단, 퇴치하는"친환경 그린 치료소"입니다. 모든 상황에 맞추어 식약허가해충퇴치제를 보다 강력하고, 안전하며 친환경적인 제품으로 맞추어 개발, 사용하고 있습니다.
        companyContactNumbers: 1800-3055
        killableBugs: 바퀴벌레, 모기, 좀벌레
        availableArea: 서울, 경기, 인천, 충청북도
        availableCounselTime: 24시간 고객센터 상담 가능
        thumbnail: http://www.hcmaster.com/images/common/top_logo.gif
        processState: 1
        engineerName: 김철수
        engineerContactNumbers: 010-1234-5678
        expectedEstimate: 100000
        visitDateTime: '2022-05-12'
...
    InsertId:
      type: object
      properties:
        insertId: 
          type: integer
          format: int64
      required: 
        - insertId
      example:
        insertId: 1
...
    ErrorApiResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
      example:
        message: 'Something went wrong. 😅'